[WIP] Add Flax LoRA Support to Dreambooth
I saw @patrickvonplaten is working on LoRA support for the non-Flax Dreambooth.
We've been taking a stab at implementing LoRA support for TPUs, taking example from
the patching method used by @cloneofsimo in cloneofsimo/lora.

I've got it successfully patching and training, but the output is currently no good.
I'm reaching the end of the time I have allocated for this—might pick it up in the future,
but for now I'm putting this up in case anyone finds it useful!